+++ src/routes/api/auth/auth.model.ts
@@ -1,0 +1,5 @@
+export interface ICurrentUser {
+    username: string;
+    password: string;
+    confirmPassword: string;
+}
