+++ src/routes/api/auth/index.json.ts
@@ -1,5 +1,6 @@
 import type { RequestHandler } from "@sveltejs/kit";
-import type { ICurrentUser, IFailureHttpResponse, ISuccessHttpResponse } from "../models";
+import type { IFailureHttpResponse, ISuccessHttpResponse } from "../base.model";
+import type { ICurrentUser } from "./auth.model";
 
 //TODO: Persist in database
 let currentUser: ICurrentUser = null;
@@ -15,7 +16,7 @@ export const get: RequestHandler = () => {
 }
 
 export const post: RequestHandler = (request): ISuccessHttpResponse<string> | IFailureHttpResponse => {
-    console.log("POST: ",  request.body);
+    console.log("POST: ", request.body);
 
     currentUser = request.body as any;
     if (currentUser) {
