+++ src/routes/api/teams/_api.json.ts
@@ -1,0 +1,4 @@
+import type { Team } from "./team.model";
+
+const teams: Team[] = [];
+
