+++ src/routes/api/teams/index.json.ts
@@ -1,0 +1,11 @@
+import type { RequestHandler } from "@sveltejs/kit";
+
+export const get: RequestHandler = () => {
+    return {}
+}
+
+export const post: RequestHandler = () => {
+    return {
+        
+    }
+}
