+++ src/routes/api/teams/team.model.ts
@@ -1,0 +1,4 @@
+export interface Team {
+    name: string;
+    members: string[]
+}
